compile unit: IO plugins, FIeldTmp, collision and ionization
