+++ include/picongpu/_defaultParam.loader
@@ -62,10 +62,8 @@
 #include "picongpu/param/collision.param"
 #include "picongpu/param/fieldSolver.param"
 #include "picongpu/param/fieldBackground.param"
-#include "picongpu/param/fileOutput.param"
 #include "picongpu/param/pngColorScales.param"
 #include "picongpu/param/png.param"
 #include "picongpu/param/particleCalorimeter.param"
-#include "picongpu/param/isaac.param"
 #include "picongpu/param/radiationObserver.param"
 // clang-format on
