+++ include/picongpu/_defaultUnitless.loader
@@ -40,8 +40,6 @@
 #include "picongpu/unitless/speciesDefinition.unitless"
 #include "picongpu/unitless/fieldBackground.unitless"
 #include "picongpu/unitless/collision.unitless"
-#include "picongpu/unitless/fileOutput.unitless"
-#include "picongpu/unitless/checkpoints.unitless"
 #include "picongpu/unitless/png.unitless"
 #include "picongpu/unitless/radiation.unitless"
 #include "picongpu/unitless/transitionRadiation.unitless"
