+++ include/picongpu/particles/filter/All.hpp
@@ -20,7 +20,11 @@
 
 #pragma once
 
-#include "picongpu/simulation_defines.hpp"
+#include "picongpu/defines.hpp"
+#include "pmacc/attribute/FunctionSpecifier.hpp"
+#include "pmacc/dimensions/DataSpace.hpp"
+
+#include <string>
 
 namespace picongpu
 {
