+++ include/picongpu/particles/filter/RelativeGlobalDomainPosition.def
@@ -19,8 +19,12 @@
 
 
 #pragma once
+
+#include "picongpu/defines.hpp"
 #include "picongpu/particles/traits/SpeciesEligibleForSolver.hpp"
 
+#include <pmacc/meta/conversion/MakeSeq.hpp>
+#include <pmacc/particles/Identifier.hpp>
 #include <pmacc/traits/HasIdentifiers.hpp>
 
 namespace picongpu
