+++ include/picongpu/particles/particleToGrid/ComputeGridValuePerFrame.hpp
@@ -23,6 +23,7 @@
 
 #include "picongpu/algorithms/Gamma.hpp"
 #include "picongpu/particles/particleToGrid/ComputeGridValuePerFrame.def"
+#include "picongpu/particles/particleToGrid/ComputeGridValuePerFrame.hpp"
 #include "picongpu/particles/particleToGrid/derivedAttributes/DerivedAttributes.hpp"
 
 #include <pmacc/math/Vector.hpp>
