+++ include/picongpu/plugins/IsaacPlugin.x.cpp
@@ -18,10 +18,12 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-#pragma once
 
 #if(ENABLE_ISAAC == 1) && (SIMDIM == DIM3)
 
+#    include "picongpu/param/fileOutput.param"
+#    include "picongpu/param/isaac.param"
+#    include "picongpu/particles/filter/filter.hpp"
 #    include "picongpu/particles/particleToGrid/ComputeFieldValue.hpp"
 #    include "picongpu/plugins/ILightweightPlugin.hpp"
 #    include "picongpu/plugins/PluginRegistry.hpp"
