+++ include/picongpu/plugins/PluginController.hpp
@@ -27,14 +27,12 @@
 #include "picongpu/particles/traits/SpeciesEligibleForSolver.hpp"
 #include "picongpu/plugins/BinEnergyParticles.hpp"
 #include "picongpu/plugins/ChargeConservation.hpp"
-#include "picongpu/plugins/Checkpoint.hpp"
 #include "picongpu/plugins/CountParticles.hpp"
 #include "picongpu/plugins/Emittance.hpp"
 #include "picongpu/plugins/EnergyFields.hpp"
 #include "picongpu/plugins/EnergyParticles.hpp"
 #include "picongpu/plugins/ILightweightPlugin.hpp"
 #include "picongpu/plugins/ISimulationPlugin.hpp"
-#include "picongpu/plugins/IsaacPlugin.hpp"
 #include "picongpu/plugins/PhaseSpace/PhaseSpace.hpp"
 #include "picongpu/plugins/PluginRegistry.hpp"
 #include "picongpu/plugins/PngPlugin.hpp"
@@ -42,7 +40,6 @@
 #include "picongpu/plugins/binning/BinningDispatcher.hpp"
 #include "picongpu/plugins/makroParticleCounter/PerSuperCell.hpp"
 #include "picongpu/plugins/multi/Master.hpp"
-#include "picongpu/plugins/openPMD/openPMDWriter.hpp"
 #include "picongpu/plugins/particleCalorimeter/ParticleCalorimeter.hpp"
 #include "picongpu/plugins/radiation/Radiation.hpp"
 #include "picongpu/plugins/radiation/VectorTypes.hpp"
