+++ share/picongpu/unit/MoveParticle.cpp
@@ -21,7 +21,6 @@
 
 #include <picongpu/simulation_defines.hpp>
 
-#include "picongpu/fields/Fields.tpp"
 #include "picongpu/particles/filter/filter.hpp"
 
 #include <algorithm>
